Internals

LazyArrays.AbstractArrayApplyStyle — Type

AbstractArrayApplyStyle

is an abstract type whose subtypes indicate how a lazy function is materialized, where the result is an AbstractArray.

LazyArrays.Add — Method

Add(A1, A2, …, AN)

A lazy representation of A1 + A2 + … + AN; i.e., a shorthand for applied(+, A1, A2, …, AN).

LazyArrays.ApplyStyle — Type

ApplyStyle

is an abstract type whose subtypes indicate how a lazy function is materialized. The default is DefaultApplyStyle which indicates that applied(f, A...) is materialized as f(A...).

LazyArrays.BroadcastLayout — Type

BroadcastLayout{F}()

is returned by MemoryLayout(A) if a matrix A is a BroadcastArray. F is the typeof function that broadcast operation is applied.

LazyArrays.DefaultApplyStyle — Type

DefaultApplyStyle

indicate that a lazy function application applied(f, A...) is materialized as f(A...).

LazyArrays.DefaultArrayApplyStyle — Type

DefaultArrayApplyStyle

is like DefaultApplyStyle but indicates that the result is an array.

LazyArrays.MulAddStyle — Type

MulAddStyle

indicates that an Applied object should be materialised via ArrayLayouts.MulAdd.

LazyArrays.MulStyle — Type

MulStyle

indicates that an Applied object should be materialized via Mul.

LazyArrays.PaddedColumns — Type

PaddedColumns{L}()

represents a vector or matrix with layout L() whose columns have been padded with zeros below, i.e., a lazy version of [A; Zeros(...)].

LazyArrays.PaddedLayout — Type

PaddedLayout{L}()

represents a matrix whose rows and columns have been padded.

LazyArrays.PaddedRows — Type

PaddedRows{L}()

represents a matrix with layout L() whose rows have been padded with zeros below, i.e., a lazy version of [A Zeros(...)].

LazyArrays.is_dotcall — Method

Check if ex is a dot-call expression like f.(x, y, z) or x .+ y .+ z.

LazyArrays.is_dotcall_nonop — Method

Check if ex is an expression like f.(x, y, z).

LazyArrays.is_dotcall_op — Method

Check if ex is an expression like x .+ y .+ z.